+++ R/dark_mode.R
@@ -7,13 +7,13 @@
 #' @param base_pal The base colour of the text. Defaults to "#bbccdd".
 #' @param title_family The font family of the title. Defaults to the base_family.
 #' @param title_face The font face of the title. Defaults to "bold".
-#' @param title_pal The colour of the title. Defaults to the base_pal.
+#' @param title_pal The colour of the title. Defaults to the base_pal first element.
 #' @param title_size The size of the title. Defaults to the base_size * 1.1.
 #' @param title_vjust The vertical adjustment of the title. Defaults to 0.
 #' @param title_margin The margin of the title. A ggplot2::margin function.
 #' @param subtitle_family The font family of the subtitle. Defaults to the base_family.
 #' @param subtitle_face The font face of the subtitle. Defaults to the base_face.
-#' @param subtitle_pal The colour of the subtitle. Defaults to the base_pal.
+#' @param subtitle_pal The colour of the subtitle. Defaults to the base_pal first element.
 #' @param subtitle_size The size of the subtitle. Defaults to the base_size.
 #' @param subtitle_vjust The vertical adjustment of the subtitle. Defaults to 1.
 #' @param subtitle_margin The margin of the subtitle. A ggplot2::margin function.
@@ -21,7 +21,7 @@
 #' @param caption_face The font face of the caption. Defaults to the base_face.
 #' @param caption_size The size of the caption. Defaults to the base_size * 0.9.
 #' @param caption_alpha The alpha of the caption pal. Defaults to 0.33. Use 1 for no alpha.
-#' @param caption_pal The colour of the caption (before caption_alpha is applied). Defaults to the base_pal.
+#' @param caption_pal The colour of the caption (before caption_alpha is applied). Defaults to the base_pal first element.
 #' @param caption_hjust The horizontal adjustment of the caption. Defaults to 0.
 #' @param caption_vjust The vertical adjustment of the caption. Defaults to 1.
 #' @param caption_margin The margin of the caption. A ggplot2::margin function.
@@ -70,7 +70,7 @@ dark_mode <- function (
     subtitle_pal = NULL,
     subtitle_size = NULL,
     subtitle_vjust = 1,
-    subtitle_margin = ggplot2::margin(t = base_size * -1, b = base_size * 2),
+    subtitle_margin = ggplot2::margin(t = base_size * -1, b = base_size + 10),
     caption_family = NULL,
     caption_face = NULL,
     caption_alpha = 0.33,
@@ -82,8 +82,8 @@ dark_mode <- function (
 ) {
 
   ggplot2::theme(
-    line = ggplot2::element_line(colour = "#bbccdd", linewidth = 10/33, linetype = 1, lineend = "square"),
-    rect = ggplot2::element_rect(fill = "#121b24", colour = "#121b24", linewidth = base_size/33, linetype = 1),
+    line = ggplot2::element_line(colour = base_pal, linewidth = 10/33, linetype = 1, lineend = "square"),
+    rect = ggplot2::element_rect(fill = pal_dark_mode[2], colour = pal_dark_mode[2], linewidth = base_size/33, linetype = 1),
     text = ggplot2::element_text(family = base_family, face = base_face, colour = base_pal, size = base_size,
                                  lineheight = 0.9, hjust = 0.5, vjust = 0.5, angle = 0, margin = ggplot2::margin(), debug = FALSE),
     axis.line = NULL,
@@ -110,7 +110,7 @@ dark_mode <- function (
     legend.spacing.x = NULL,
     legend.spacing.y = NULL,
     legend.margin = ggplot2::margin(),
-    legend.key = ggplot2::element_rect(fill = "#1f2f3e", colour = "#1f2f3e"),
+    legend.key = ggplot2::element_rect(fill = pal_dark_mode[3], colour = pal_dark_mode[3]),
     legend.key.size = grid::unit(1.2, "lines"),
     legend.key.height = NULL,
     legend.key.width = NULL,
@@ -125,10 +125,10 @@ dark_mode <- function (
     legend.box.margin = ggplot2::margin(l = base_size * 0.5),
     legend.box.background = NULL,
     legend.box.spacing = NULL,
-    panel.background = ggplot2::element_rect(fill = "#1f2f3e", colour = "#1f2f3e"),
+    panel.background = ggplot2::element_rect(fill = pal_dark_mode[3], colour = pal_dark_mode[3]),
     panel.border = ggplot2::element_blank(),
     panel.grid = NULL,
-    panel.grid.major = ggplot2::element_line(colour = "#2C3A48", linewidth = ggplot2::rel(0.5)),
+    panel.grid.major = ggplot2::element_line(colour = pal_dark_mode[4], linewidth = ggplot2::rel(0.5)),
     panel.grid.minor = ggplot2::element_blank(),
     panel.spacing = grid::unit(1.25, "lines"),
     panel.spacing.x = NULL,