+++ R/gg_text.R
@@ -83,6 +83,7 @@
 #'     y = mpg,
 #'     col = mpg,
 #'     label = model,
+#'     size = 3.53,
 #'     y_include = c(min(.$mpg) * 0.95, max(.$mpg) * 1.05),
 #'     pal = RColorBrewer::brewer.pal(9, "RdBu"),
 #'     y_title = "Miles per gallon"
@@ -271,7 +272,7 @@ gg_text <- function(
       dplyr::mutate(dplyr::across(!!col, function(x) factor(x)))
   }
 
-  if (flipped & col_character | col_factor) {
+  if (flipped & (col_character | col_factor)) {
     data <- data %>%
       dplyr::mutate(dplyr::across(!!col, function(x) forcats::fct_rev(x)))
   }