+++ NEWS.md
@@ -1,3 +1,8 @@
+# ggblanket 5.1.0
+
+* Fixed bug where `col` variable was reversing for non-flipped.
+* Added `pal_light_mode` and `pal_dark_mode`.
+
 # ggblanket 5.0.0
 
 * Added `mapping` argument to allow extra aesthetics to be included, such as shape (or use delayed evaluation for aesthetics other than `col`, `colour`, `fill` or `alpha`).
@@ -8,7 +13,7 @@
 * Breaking: removed `lower`, `middle`, `upper`, `xlower`, `xmiddle`, and `xupper` arguments from `gg_blank`.
 * Updated `*_breaks` defaults.
 * Made plot order logical variables with `TRUE` first.  
-* Added `pal_discrete2` with 6 colour blind safe colours for a light background theme.
+* Added `pal_discrete2` with 6 colour blind safe colours for a light background theme. 
 
 # ggblanket 4.0.0